+++ CHANGELOG.md
@@ -10,6 +10,9 @@ Breaking Changes:
 
 - The mpack configuration `mpack-config.h` file is now optional, and requires `MPACK_HAS_CONFIG` in order to be included. This means you must define `MPACK_HAS_CONFIG` when upgrading or your config file will be ignored! (It is recommended to delete your config file and use the defaults.)
 
+Changes:
+
+- The reader's skip function is no longer ignored under `MPACK_OPTIMIZE_FOR_SIZE`.
 
 
 Release Versions